refactor(backend): move structure domain into farm/structure
Move LocationViewSet, FieldViewSet, BedViewSet and
BedLayoutByLocationView plus the Location/Field/Bed/layout serializers
and the duplicate-name message constants into a farm/structure domain
package. Route table unchanged; test imports repointed.
